#62 refactor: 패키지 명을 application으로 변경
