+++ backend/src/test/java/meetween/backend/authentication/application/client/StubOAuthClient.java
@@ -1,4 +1,4 @@
-package meetween.backend.authentication.service.client;
+package meetween.backend.authentication.application.client;
 
 import meetween.backend.authentication.domain.client.OAuthClient;
 import meetween.backend.authentication.domain.oauthmember.KakaoOAuthMember;
