+++ backend/src/test/java/meetween/backend/config/TestConfig.java
@@ -1,6 +1,6 @@
 package meetween.backend.config;
 
-import meetween.backend.authentication.service.client.StubOAuthClient;
+import meetween.backend.authentication.application.client.StubOAuthClient;
 import meetween.backend.authentication.domain.client.OAuthClient;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
