+++ backend/src/test/java/meetween/backend/support/annotation/ControllerTest.java
@@ -2,18 +2,17 @@ package meetween.backend.support.annotation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import meetween.backend.appointment.presentation.AppointmentController;
-import meetween.backend.appointment.service.AppointmentService;
-import meetween.backend.appointment.service.AppointmentUserService;
+import meetween.backend.appointment.application.AppointmentService;
 import meetween.backend.authentication.infrastructure.jwt.JwtTokenProvider;
 import meetween.backend.authentication.infrastructure.uri.OAuthUriProvider;
 import meetween.backend.authentication.presentataion.AuthArgumentResolver;
 import meetween.backend.authentication.presentataion.AuthController;
 import meetween.backend.authentication.presentataion.BearerTokenExtractor;
-import meetween.backend.authentication.service.AuthService;
+import meetween.backend.authentication.application.AuthService;
 import meetween.backend.category.presentation.CategoryController;
-import meetween.backend.category.service.CategoryService;
+import meetween.backend.category.application.CategoryService;
 import meetween.backend.member.presentation.MemberController;
-import meetween.backend.member.service.MemberService;
+import meetween.backend.member.application.MemberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -57,9 +56,6 @@ public abstract class ControllerTest {
 
     @MockBean
     protected CategoryService categoryService;
-
-    @MockBean
-    protected AppointmentUserService appointmentUserService;
 
     @MockBean
     protected OAuthUriProvider oAuthUriProvider;
