+++ backend/src/test/java/meetween/backend/support/annotation/ServiceTest.java
@@ -1,6 +1,6 @@
 package meetween.backend.support.annotation;
 
-import meetween.backend.authentication.service.AuthService;
+import meetween.backend.authentication.application.AuthService;
 import meetween.backend.config.TestConfig;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
